+++ neural_network_api/optimiser_kernel.h
@@ -14,20 +14,20 @@
 void adam_update_momentum(
 	float * d_momentum, 
 	float * d_derivatives, 
-	float beta, 
+	double beta, 
 	size_t size);
 void adam_update_velocity(
 	float * d_velocity, 
 	float * d_derivatives, 
-	float beta, 
+	double beta,
 	size_t size);
 void adam_update_parameters(
 	float * d_params, 
 	float * d_momentum, 
 	float * d_velocity, 
-	float learning_rate, 
-	float beta1, 
-	float beta2, 
-	float epsilon,
+	double learning_rate,
+	double beta1,
+	double beta2,
+	double epsilon,
 	int t_step, 
 	size_t size);
